move strategy/config to own directory
