+++ shibbo/src/server.ts
@@ -5,22 +5,15 @@ import express, {
 } from "express"
 import morgan from "morgan"
 import passport from "passport"
-import { MultiSamlStrategy } from "passport-saml"
 import shibbolethCharsetMiddleware from "unfuck-utf8-headers-middleware"
 
 import {
   DOMAIN,
-  HAKA_IDP_URL,
-  HY_IDP_URL,
-  MOOCFI_CERTIFICATE,
   PORT,
   SHIBBOLETH_HEADERS,
 } from "./config"
 import { callbackHandler } from "./handlers"
-import {
-  convertObjectKeysToLowerCase,
-  encodeRelayState,
-} from "./util"
+import { createSamlStrategy } from "./saml"
 
 const isProduction = process.env.NODE_ENV === "production"
 
@@ -63,11 +56,10 @@ app.use((req, res, next) => {
 app.use(passport.initialize())
 
 
-const providers: Record<string, string> = {
-  hy: HY_IDP_URL,
-  haka: HAKA_IDP_URL,
-}
-
+passport.use(
+  "hy-haka",
+  createSamlStrategy()
+)
 /*passport.serializeUser((user, done) => {
   console.log("serialize", user)
   done(null, user)
@@ -78,41 +70,6 @@ passport.deserializeUser((user, done) => {
   done(null, user as any)
 })*/
 
-passport.use(
-  "hy-haka",
-  new MultiSamlStrategy(
-    {
-      passReqToCallback: true,
-      getSamlOptions(req, done) {
-        const relayState = encodeRelayState(req)
-        const { provider, action } = req.params
-        const language = req.query.language || req.params.language || "en"
-
-        done(null, {
-          name: "multi",
-          path: `/callbacks/${provider}/${action}/${language}`,
-          entryPoint: providers[provider],
-          issuer: "https://mooc.fi/sp",
-          cert: MOOCFI_CERTIFICATE,
-          authnRequestBinding: "HTTP-POST",
-          forceAuthn: true,
-          identifierFormat:
-            "urn:oasis:names:tc:SAML:2.0:nameid-format:transient",
-          additionalParams: {
-            RelayState: relayState,
-          },
-        })
-      },
-    },
-    (_req, profile: any, done) => {
-      console.log("got profile", profile)
-      done(
-        null,
-        convertObjectKeysToLowerCase(profile?.attributes)
-      )
-    },
-  ),
-)
 
 app.get("/:action/:provider", callbackHandler)
